added edit page
Took 1 hour 28 minutes
