+++ Electron-Frontend/main.js
@@ -1,29 +1,22 @@
-const {app, BrowserWindow, Menu,ipcMain} = require('electron')
+const {app, BrowserWindow, Menu} = require('electron')
 const shell = require('electron').shell
 
-let mainWindow
-let loginWindow
-
 function createWindow() {
-    mainWindow = new BrowserWindow({
+    const mainWindow = new BrowserWindow({
         width: 800,
         height: 600,
-        show: false,
+        show: true,
         webPreferences: {
             nodeIntegration: true
         }
     })
     mainWindow.loadFile('src/index.html')
 
-    loginWindow = new BrowserWindow({
+    const loginWindow = new BrowserWindow({
         parent: mainWindow,
         width: 400,
         height: 300,
-        frame: false,
-        resizable: false,
-        webPreferences: {
-            nodeIntegration: true
-        }
+        frame: false
     })
     loginWindow.loadFile('src/login.html')
 
@@ -52,13 +45,6 @@ function createWindow() {
     Menu.setApplicationMenu(menu);
 
 }
-
-ipcMain.on('entry-accepted', (event, arg) =>{
-    if(arg === 'true'){
-        mainWindow.show()
-        loginWindow.hide()
-    }
-})
 
 app.whenReady().then(createWindow)
