+++ Electron-Frontend/src/index.html
@@ -5,8 +5,8 @@
     <title>Hello World</title>
     <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline';"/>
     <link rel="stylesheet" href="./contents/bootstrap/4.5.3/css/bootstrap.min.css" />
-    <script src="contents/jqueryData/3.5.1/jquery-3.5.1.min.js"></script>
-    <script src="contents/popper/1.16.3/popper.min.js"></script>
+    <script src="./contents/jquery/3.5.1/jquery-3.5.1.min.js"></script>
+    <script src="./contents/popper.js/1.16.3/popper.min.js"></script>
     <script src="./contents/bootstrap/4.5.3/js/bootstrap.min.js"></script>
     <link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/bootstrap/3.4.1/css/bootstrap.min.css">
     <script src="https://ajax.googleapis.com/ajax/libs/jquery/3.5.1/jquery.min.js"></script>
@@ -33,6 +33,11 @@
         <br/>
         <p>Beschreibung: <p id="description"></p></p>
     </div>
+
+
+    We are using node <script>document.write(process.versions.node)</script>,
+    Chrome <script>document.write(process.versions.chrome)</script>
+    and Electron <script>document.write(process.versions.electron)</script>.
 </body>
 </html>
